public-split-he-h: Pulled out he_internal.h and conn_internal.h from he.h and conn.h
