--- python/prod/he.h.footer
@@ -1,1 +1,0 @@
-#endif
