+++ src/he/flow.c
@@ -21,6 +21,7 @@
 #include "core.h"
 #include "msg_handlers.h"
 #include "conn.h"
+#include "conn_internal.h"
 #include "plugin_chain.h"
 
 #ifndef WOLFSSL_USER_SETTINGS
